refactor: Portal에서 클릭 이벤트 훅으로 분리
